ON lrs_convert: [1] removed datasets from point_datasets class variable where required; [2] changed name to just lrs_convert and deleted original script; [3] added `spatial key` when merging schemas in helpers.export; [4] added yaml update task to processing_task.md.
